\name{FdPot-package}
\alias{FdPot-package}
\alias{FdPot}
\docType{package}
\title{
\packageTitle{FdPot}
}
\description{
\packageDescription{FdPot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FdPot}
\packageIndices{FdPot}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{FdPot}

Maintainer: \packageMaintainer{FdPot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
