% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFdorct.R
\name{pFdorct}
\alias{pFdorct}
\title{Build and fit an FD-POT that classifies functional data}
\usage{
pFdorct(
  y,
  X,
  basis.degree,
  depth = 2,
  alpha = 0.5,
  similarity.method = "d0.L2",
  n.solve = 20,
  m_cost = 0.5,
  gamma = 512,
  seed = 21071865
)
}
\arguments{
\item{y}{vector of integers with the labels. They MUST be numbered starting from 0}

\item{X}{the smoothed functional dats object of class fdSmooth}

\item{depth}{the classificatin tree depth}

\item{alpha}{the weight given in the objective function (stronger alpha, higher penalty for dissimilarity in each leaf node)}

\item{similarity.method}{the method used to obtain the dissimilarity between two functional data. Currently only L2 norm of the difference supported}

\item{n.solve}{how many optimisations to carry out (each from a different starting point)}

\item{m_cost}{the misclassification cost (best left at default value)}

\item{gamma}{the randomisation factor (best left at default value)}

\item{seed}{the random seed}
}
\description{
Given a vector of labels and an fdSmooth object (smoothed functional data), builds and fits a penalised optimal randomised decision tree
}
