% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pFdorct_Rcpp}
\alias{pFdorct_Rcpp}
\title{Build and fit an FD-classification penalised tree}
\usage{
pFdorct_Rcpp(
  y,
  X_coeffs,
  X_argvals,
  X_basis_df,
  X_basis_degree,
  basis_type = "BSpline",
  depth = 2L,
  alpha = 0.1,
  similarity_method = "d0.L2",
  n_feats = 10L,
  n_solve = 20L,
  gamma = 512,
  seed = 41703192L
)
}
\arguments{
\item{y}{labels vector. Note they have to be integers starting from 0}

\item{X_coeffs}{p x n matrix with the coefficients fitted in the smoothing (see the examples of the library), where p is the number of coeffients for func. datum and n the sample size}

\item{X_argvals}{vector of lower and upper bounds of the domain}

\item{X_basis_df}{the degrees of freedom of the basis}

\item{X_basis_degree}{the degree of the (b-spline) basis}

\item{basis_type}{the basis type string, BSpline is the only one currently supported}

\item{depth}{the tree depth. Make}

\item{alpha}{the hyperparameter for the penalty in the objective function. Higher alpha, higher weight for the penalty}

\item{n_feats}{how features to use at each node of the tree to perform a split. An equal-length partition of the size of n_feats is created; each feat is the integral of the func. datum in a set that is part of the pariition}

\item{n_solve}{how many different trees to fit starting from different init points}

\item{gamma}{the randomisation factor for the ORCT, best kept default}

\item{seed}{random seed for reproducibility}
}
\description{
instantiates and fits a Functional Data Penalised Optimial Randomised Decision Tree
}
\note{
The s3 class p.fdorct provides an interface for this method
}
